+++ src/backend/access/rmgrdesc/heapdesc.c
@@ -78,8 +78,8 @@ plan_elem_desc(StringInfo buf, void *plan, void *data)
 	xlhp_freeze_plan *new_plan = (xlhp_freeze_plan *) plan;
 	OffsetNumber **offsets = data;
 
-	appendStringInfo(buf, "{ xmax: %llu, infomask: %u, infomask2: %u, ntuples: %u",
-					 (unsigned long long) new_plan->xmax,
+	appendStringInfo(buf, "{ xmax: %" PRIu64 ", infomask: %u, infomask2: %u, ntuples: %u",
+					 new_plan->xmax,
 					 new_plan->t_infomask, new_plan->t_infomask2,
 					 new_plan->ntuples);
 
@@ -199,8 +199,8 @@ heap_desc(StringInfo buf, XLogReaderState *record)
 	{
 		xl_heap_delete *xlrec = (xl_heap_delete *) rec;
 
-		appendStringInfo(buf, "xmax: %llu, off: %u, ",
-						 (unsigned long long) xlrec->xmax, xlrec->offnum);
+		appendStringInfo(buf, "xmax: %" PRIu64 ", off: %u, ",
+						 xlrec->xmax, xlrec->offnum);
 		infobits_desc(buf, xlrec->infobits_set, "infobits");
 		appendStringInfo(buf, ", flags: 0x%02X", xlrec->flags);
 	}
@@ -208,26 +208,21 @@ heap_desc(StringInfo buf, XLogReaderState *record)
 	{
 		xl_heap_update *xlrec = (xl_heap_update *) rec;
 
-		appendStringInfo(buf, "old_xmax: %llu, old_off: %u, ",
-						 (unsigned long long) xlrec->old_xmax,
-						 xlrec->old_offnum);
+		appendStringInfo(buf, "old_xmax: %" PRIu64 ", old_off: %u, ",
+						 xlrec->old_xmax, xlrec->old_offnum);
 		infobits_desc(buf, xlrec->old_infobits_set, "old_infobits");
-		appendStringInfo(buf, ", flags: 0x%02X, new_xmax: %llu, new_off: %u",
-						 xlrec->flags,
-						 (unsigned long long) xlrec->new_xmax,
-						 xlrec->new_offnum);
+		appendStringInfo(buf, ", flags: 0x%02X, new_xmax: %" PRIu64 ", new_off: %u",
+						 xlrec->flags, xlrec->new_xmax, xlrec->new_offnum);
 	}
 	else if (info == XLOG_HEAP_HOT_UPDATE)
 	{
 		xl_heap_update *xlrec = (xl_heap_update *) rec;
 
-		appendStringInfo(buf, "old_xmax: %llu, old_off: %u, ",
-						 (unsigned long long) xlrec->old_xmax,
-						 xlrec->old_offnum);
+		appendStringInfo(buf, "old_xmax: %" PRIu64 ", old_off: %u, ",
+						 xlrec->old_xmax, xlrec->old_offnum);
 		infobits_desc(buf, xlrec->old_infobits_set, "old_infobits");
-		appendStringInfo(buf, ", flags: 0x%02X, new_xmax: %llu, new_off: %u",
-						 xlrec->flags, (unsigned long long) xlrec->new_xmax,
-						 xlrec->new_offnum);
+		appendStringInfo(buf, ", flags: 0x%02X, new_xmax: %" PRIu64 ", new_off: %u",
+						 xlrec->flags, xlrec->new_xmax, xlrec->new_offnum);
 	}
 	else if (info == XLOG_HEAP_TRUNCATE)
 	{
@@ -249,8 +244,8 @@ heap_desc(StringInfo buf, XLogReaderState *record)
 	{
 		xl_heap_lock *xlrec = (xl_heap_lock *) rec;
 
-		appendStringInfo(buf, "xmax: %llu, off: %u, ",
-						 (unsigned long long) xlrec->xmax, xlrec->offnum);
+		appendStringInfo(buf, "xmax: %" PRIu64 ", off: %u, ",
+						 xlrec->xmax, xlrec->offnum);
 		infobits_desc(buf, xlrec->infobits_set, "infobits");
 		appendStringInfo(buf, ", flags: 0x%02X", xlrec->flags);
 	}
@@ -284,8 +279,8 @@ heap2_desc(StringInfo buf, XLogReaderState *record)
 
 			memcpy(&conflict_xid, rec + SizeOfHeapPrune, sizeof(TransactionId));
 
-			appendStringInfo(buf, "snapshotConflictHorizon: %llu",
-							 (unsigned long long) conflict_xid);
+			appendStringInfo(buf, "snapshotConflictHorizon: %" PRIu64,
+							 conflict_xid);
 		}
 
 		appendStringInfo(buf, ", isCatalogRel: %c",
@@ -348,9 +343,8 @@ heap2_desc(StringInfo buf, XLogReaderState *record)
 	{
 		xl_heap_visible *xlrec = (xl_heap_visible *) rec;
 
-		appendStringInfo(buf, "snapshotConflictHorizon: %llu, flags: 0x%02X",
-						 (unsigned long long) xlrec->snapshotConflictHorizon,
-						 xlrec->flags);
+		appendStringInfo(buf, "snapshotConflictHorizon: %" PRIu64 ", flags: 0x%02X",
+						 xlrec->snapshotConflictHorizon, xlrec->flags);
 	}
 	else if (info == XLOG_HEAP2_MULTI_INSERT)
 	{
@@ -371,8 +365,8 @@ heap2_desc(StringInfo buf, XLogReaderState *record)
 	{
 		xl_heap_lock_updated *xlrec = (xl_heap_lock_updated *) rec;
 
-		appendStringInfo(buf, "xmax: %llu, off: %u, ",
-						 (unsigned long long) xlrec->xmax, xlrec->offnum);
+		appendStringInfo(buf, "xmax: %" PRIu64 ", off: %u, ",
+						 xlrec->xmax, xlrec->offnum);
 		infobits_desc(buf, xlrec->infobits_set, "infobits");
 		appendStringInfo(buf, ", flags: 0x%02X", xlrec->flags);
 	}
@@ -402,9 +396,9 @@ heap3_desc(StringInfo buf, XLogReaderState *record)
 	{
 		xl_heap_base_shift *xlrec = (xl_heap_base_shift *) rec;
 
-		appendStringInfo(buf, "%s delta %lld ",
+		appendStringInfo(buf, "%s delta %" PRIu64 " ",
 						 xlrec->multi ? "MultiXactId" : "XactId",
-						 (long long) xlrec->delta);
+						 xlrec->delta);
 	}
 }
