+++ src/backend/access/transam/xact.c
@@ -1790,8 +1790,8 @@ RecordTransactionAbort(bool isSubXact)
 	 * Check that we haven't aborted halfway through RecordTransactionCommit.
 	 */
 	if (TransactionIdDidCommit(xid))
-		elog(PANIC, "cannot abort transaction %llu, it was already committed",
-			 (unsigned long long) xid);
+		elog(PANIC, "cannot abort transaction %" PRIu64 ", it was already committed",
+			 xid);
 
 	/*
 	 * Are we using the replication origins feature?  Or, in other words, are
@@ -5681,20 +5681,18 @@ ShowTransactionStateRec(const char *str, TransactionState s)
 	{
 		int			i;
 
-		appendStringInfo(&buf, ", children: %llu",
-						 (unsigned long long) s->childXids[0]);
+		appendStringInfo(&buf, ", children: %" PRIu64, s->childXids[0]);
 		for (i = 1; i < s->nChildXids; i++)
-			appendStringInfo(&buf, " %llu",
-							 (unsigned long long) s->childXids[i]);
+			appendStringInfo(&buf, " %" PRIu64, s->childXids[i]);
 	}
 	ereport(DEBUG5,
-			(errmsg_internal("%s(%d) name: %s; blockState: %s; state: %s, xid/subid/cid: %llu/%llu/%u%s%s",
+			(errmsg_internal("%s(%d) name: %s; blockState: %s; state: %s, xid/subid/cid: %" PRIu64 "/%" PRIu64 "/%u%s%s",
 							 str, s->nestingLevel,
 							 PointerIsValid(s->name) ? s->name : "unnamed",
 							 BlockStateAsString(s->blockState),
 							 TransStateAsString(s->state),
-							 (unsigned long long) XidFromFullTransactionId(s->fullTransactionId),
-							 (unsigned long long) s->subTransactionId,
+							 XidFromFullTransactionId(s->fullTransactionId),
+							 s->subTransactionId,
 							 (unsigned int) currentCommandId,
 							 currentCommandIdUsed ? " (used)" : "",
 							 buf.data)));
