+++ src/backend/replication/logical/logical.c
@@ -1723,8 +1723,8 @@ LogicalIncreaseXminForSlot(XLogRecPtr current_lsn, TransactionId xmin)
 	SpinLockRelease(&slot->mutex);
 
 	if (got_new_xmin)
-		elog(DEBUG1, "got new catalog xmin %llu at %X/%X",
-			 (unsigned long long) xmin, LSN_FORMAT_ARGS(current_lsn));
+		elog(DEBUG1, "got new catalog xmin %" PRIu64 " at %X/%X",
+			 xmin, LSN_FORMAT_ARGS(current_lsn));
 
 	/* candidate already valid with the current flush position, apply */
 	if (updated_xmin)
