+++ src/backend/replication/pgoutput/pgoutput.c
@@ -619,8 +619,8 @@ pgoutput_commit_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn,
 
 	if (!sent_begin_txn)
 	{
-		elog(DEBUG1, "skipped replication of an empty transaction with XID: %llu",
-			 (unsigned long long) txn->xid);
+		elog(DEBUG1, "skipped replication of an empty transaction with XID: %" PRIu64,
+			 txn->xid);
 		return;
 	}
